requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
mysqlclient==2.1.1
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
uWSGI==2.0.21
